Move script approval related tests to a separate test class
